autotester-package: reorganize entire package structure. WARNING: THIS BREAKS EVERYTHING
