+++ autotester/autotester/exceptions/__init__.py
@@ -1,0 +1,6 @@
+""" 
+Custom Exception Type for use in MarkUs
+"""
+
+class MarkUsError(Exception):
+    pass
